MeowBowWow2emoji/requirements.txt
asgiref==3.5.0
Django==3.2.13
pytz==2022.1
sqlparse==0.4.2
typing_extensions==4.1.1
uWSGI==2.0.20
tensorflow-gpu==1.14
django-cors-headers==3.11.0
djangorestframework==3.13.1
Pillow==9.1.0

